requirements.txt
flask==2.3.2
sqlalchemy==2.0.19
pydantic==1.10.12
python-dotenv==1.0.0
pytest==7.4.2
pre-commit==3.4.0
black==24.1.0
isort==5.12.0
flake8==6.1.0
requests